WIP reimpl of Pot
Not done and won't compile. But need to save my work before big edit.

This has a Pot that contains Vec<InnerPot>, and when handling bets it
tries to idx into the correct InnerPot.

Plenty of stuff is still broken (all the tests) and it won't even
compile because bet_helper as issues with my mutable borrows.

Also I have not yet added sorting of the working pots such that the ones
with the lowest max_in are at the front and the ones with the largest
(or none) at the back. I believe this is necessary.