Still WIP ... issue when overflowing.
I'm saving my changes here breifly as I believe I'm running into a
problem that Richard already solved, and I'm not sure all my changes are
necessary and worth it. It's time to see if I can pull in his ideas for
addressing this, or if I should roll all the way back and just implement
the original motivation for this branch on top of his code.

    The problem that I think he already solved:

Say there's two pots already. Player 1 (p1) is getting inserted into the
first. It's a pot with a cap and he's betting more than the cap. So his
remaining bet needs to roll over into the second. The second pot he is
already in. What the code does right now is replace his current entry,
but the right thing to do is to add to his current entry. An issue may
arise, however, if the second pot is capped and adding to his current
entry takes him too high. Will recursion solve it cleanly?

    This original motivation for this branch:

Issue #31. The core idea in this branch that may be capable of pulling
out and ipmlementing on top of Richard's code is saving a player's AllIn
status along with their wager. Instead of HashMap<PlayerId, Currency>,
it's HashMap<PlayerId, Stake>, where Stake is (is_allin, amount). This
alone (w/o the rest of this branch) I believe would address issue #31.