Remove MinRaiseSet, combine with CurrentBetSet
These events are often (but not always) logged together because they
often change at the same time.

Now CurrentBetSet contains both the (maybe new) current_bet to match and
the (maybe new) min_raise to match. If it is logged, one or both
changed, but it is not stated which.

Closes #43